Attempt to add unit test. Note nunit doesnt work because of dll versions, and xunit can't load the assembly
