+++ Samples/Aqovia.PactProducerVerifier.Sample.Test/TestStateProvider.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 
 namespace Aqovia.PactProducerVerifier.Sample.Test
