+++ Samples/Aqovia.PactProducerVerifier.Sample/Aqovia.PactProducerVerifier.Sample.csproj
@@ -1,15 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net462</TargetFramework>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
+    <TargetFramework>net462</TargetFramework>    
     <PackageTargetFallback>$(PackageTargetFallback);portable-win+net45+wp8+win81+wpa8</PackageTargetFallback>
     <IsPackable>false</IsPackable>
     <Platforms>AnyCPU;x64</Platforms>    
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|net462|AnyCPU'">
-    <PlatformTarget>x64</PlatformTarget>
+    <PlatformTarget>x86</PlatformTarget>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|net462|x64'">
@@ -20,6 +19,7 @@
     <PackageReference Include="GitInfo" Version="2.0.6" />
     <PackageReference Include="NUnit" Version="3.9.0" />
     <PackageReference Include="uHttpSharp" Version="0.1.6.22" />
+    <PackageReference Include="xunit" Version="2.3.1" />
   </ItemGroup>  
 
   <ItemGroup>
